Updates after conflicts

+++ Backend/Models/ReviewDTO.cs
@@ -4,7 +4,7 @@
     {
         public int UserId { get; set; }
         public int ProductId { get; set; }
-        public decimal Rating { get; set; }
+        public int Rating { get; set; }
         public string Comment { get; set; }
     }
 }
